- Create repository pattern of test type data
