+++ app/src/main/java/edu/upc/fib/pes_infovid19/domain/TestTypeRepository.kt
@@ -1,0 +1,11 @@
+package edu.upc.fib.pes_infovid19.domain
+
+import androidx.lifecycle.LiveData
+import edu.upc.fib.pes_infovid19.ui.main.TestType
+
+interface TestTypeRepository {
+    fun getTestTypes(): LiveData<List<TestType>>
+    fun removeTestType(id: String)
+    fun modifyTestType(id: String, testType: TestType)
+    fun createTestType(testType: TestType)
+}
